+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Added first 300 bytes of detection results as `Finding.details` when the
+  detection represents a vulnerability in the set of `vmdrFindingResultQids`
+  provided in the configuration. This is optional and it will come at a
+  significant processing cost due to the number of additional bytes transferred
+  for all host detections.
+
 ## [5.9.1] - 2021-12-01
 
 ### Changed